refactor : Location 추출
