+++ src/main/java/com/example/demo/context/Location.java
@@ -1,0 +1,10 @@
+package com.example.demo.context;
+
+public record Location(
+        Rank rank,
+        File file
+) {
+    public boolean isValid() {
+        return rank != null && file != null;
+    }
+}
