+++ src/main/java/com/example/demo/piece/Piece.java
@@ -1,7 +1,7 @@
 package com.example.demo.piece;
 
-import com.example.demo.context.Board;
 import com.example.demo.context.File;
+import com.example.demo.context.Location;
 import com.example.demo.context.Rank;
 
 public abstract class Piece implements Comparable<Piece> {
@@ -76,8 +76,8 @@ public abstract class Piece implements Comparable<Piece> {
         return this.rank == rank && this.file == file;
     }
 
-    public Board.Location getLocation(){
-        return new Board.Location(this.rank, this.file);
+    public Location getLocation(){
+        return new Location(this.rank, this.file);
     }
 
     public File getFile() {
