+++ src/test/java/com/example/demo/context/GameTest.java
@@ -16,7 +16,6 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Stream;
 
-import static com.example.demo.context.Board.Location;
 import static com.example.demo.context.Board.createBoard;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
